Fix ember-modifier v3.2 compatibility compatibility
Condition about whether to use new or old version of `on-key` modifier class was flipped.

Additionally, this adds `useWorkspaces` to ember-try config
so that proper version of `ember-modifier` is used
across the monorepo.

This is needed so that `ember-modifier-X` ember-try scenarios
resolve to the same version of `ember-modifier` both in the
`test-app` and `addon` workspaces.

`ember-classic` scenario should not use `useWorkspaces` flag as only test-app need to get classic flags.